+++ design-toolkit/components/src/components/checkbox/styles.ts
@@ -36,6 +36,7 @@ export const CheckboxStyles = tv({
       'group-enabled/checkbox:group-selected/checkbox:group-focus/checkbox:outline-interactive-hover',
       'group-enabled/checkbox:group-selected/checkbox:group-hover/checkbox:outline-interactive-hover',
       'group-disabled/checkbox:outline-interactive-disabled',
+      'group-disabled/checkbox:group-indeterminate/checkbox:fg-inverse-light group-disabled/checkbox:group-indeterminate/checkbox:bg-interactive-disabled',
       'group-disabled/checkbox:group-selected/checkbox:fg-inverse-light group-disabled/checkbox:group-selected/checkbox:bg-interactive-disabled',
     ],
     label: [