+++ design-toolkit/components/src/index.ts
@@ -19,23 +19,20 @@ export {
   AccordionStyles,
   AccordionStylesDefaults,
 } from './components/accordion/styles';
-export type { AccordionStyleVariants } from './components/accordion/styles';
 export type {
   AccordionGroupProps,
   AccordionHeaderProps,
   AccordionPanelProps,
   AccordionProps,
+  AccordionStyleVariants,
   AccordionTriggerProps,
 } from './components/accordion/types';
 export { Avatar, AvatarContext } from './components/avatar';
 export { AvatarStyles } from './components/avatar/styles';
-export type {
-  AvatarProps,
-  AvatarProviderProps,
-} from './components/avatar/types';
+export type { AvatarProps } from './components/avatar/types';
 export { Badge, BadgeContext } from './components/badge';
 export { BadgeStyles, BadgeStylesDefaults } from './components/badge/styles';
-export type { BadgeProps, BadgeProviderProps } from './components/badge/types';
+export type { BadgeProps } from './components/badge/types';
 export { Box } from './components/box';
 export type { BoxProps } from './components/box';
 export {
@@ -54,12 +51,9 @@ export {
 } from './components/button/styles';
 export type {
   ButtonProps,
-  ButtonProviderProps,
   ButtonStyleVariants,
   LinkButtonProps,
-  LinkButtonProviderProps,
   ToggleButtonProps,
-  ToggleButtonProviderProps,
 } from './components/button/types';
 export {
   Checkbox,
@@ -83,7 +77,6 @@ export type {
   BaseChipProps,
   ChipListProps,
   ChipProps,
-  ChipProviderProps,
   DeletableChipProps,
   SelectableChipProps,
 } from './components/chip/types';
@@ -97,7 +90,6 @@ export {
 } from './components/classification-badge/styles';
 export type {
   ClassificationBadgeProps,
-  ClassificationBadgeProviderProps,
   ClassificationBadgeStyleVariants,
 } from './components/classification-badge/types';
 export {
@@ -110,7 +102,6 @@ export {
 } from './components/classification-banner/styles';
 export type {
   ClassificationBannerProps,
-  ClassificationBannerProviderProps,
   ClassificationBannerStyleVariants,
 } from './components/classification-banner/types';
 export { ColorPicker, useColorPickerState } from './components/color-picker';
@@ -119,26 +110,31 @@ export { ComboBox, ComboBoxSection } from './components/combobox';
 export type { ComboBoxProps } from './components/combobox';
 export { DateField } from './components/date-field';
 export type { DateFieldProps } from './components/date-field';
+export { DetailsList, DetailsListContext } from './components/details-list';
+export {
+  DetailsListStyles,
+  DetailsListStylesDefaults,
+} from './components/details-list/styles';
+export type {
+  DetailsListLabelProps,
+  DetailsListProps,
+  DetailsListValueProps,
+} from './components/details-list/types';
 export { Dialog, DialogTitle } from './components/dialog';
 export type { DialogProps } from './components/dialog';
+export { Hero, HeroContext } from './components/hero';
+export { HeroStyles } from './components/hero/styles';
+export type { HeroProps } from './components/hero/types';
 export { Icon, IconContext } from './components/icon';
 export { IconStyles } from './components/icon/styles';
-export type { IconProps, IconProviderProps } from './components/icon/types';
+export type { IconProps } from './components/icon/types';
 export { Input, InputContext } from './components/input';
 export { InputStyles, InputStylesDefaults } from './components/input/styles';
 export type { InputProps } from './components/input/types';
 export { Label, LabelContext } from './components/label';
 export { LabelStyles } from './components/label/styles';
-export type { LabelProps, LabelProviderProps } from './components/label/types';
-export {
-  Menu,
-  MenuContext,
-  MenuDescription,
-  MenuItem,
-  MenuLabel,
-  MenuSection,
-  MenuSeparator,
-} from './components/menu';
+export type { LabelProps } from './components/label/types';
+export { Menu, MenuContext } from './components/menu';
 export { MenuStyles, MenuStylesDefaults } from './components/menu/styles';
 export type { MenuStyleVariants } from './components/menu/styles';
 export type {
@@ -186,21 +182,23 @@ export type {
 } from './components/query-builder';
 export { Radio, RadioContext } from './components/radio';
 export { RadioStyles } from './components/radio/styles';
-export type {
-  RadioGroupProps,
-  RadioProps,
-  RadioProviderProps,
-} from './components/radio/types';
+export type { RadioGroupProps, RadioProps } from './components/radio/types';
 export { SearchField } from './components/search-field';
 export type { SearchFieldProps } from './components/search-field';
+export { Skeleton } from './components/skeleton';
+export {
+  SkeletonStyles,
+  SkeletonStylesDefaults,
+} from './components/skeleton/styles';
+export type {
+  SkeletonProps,
+  SkeletonStyleVariants,
+} from './components/skeleton/types';
 export { RangeSlider, Slider } from './components/slider';
 export type { RangeSliderProps, SliderProps } from './components/slider';
 export { Switch, SwitchContext } from './components/switch';
 export { SwitchStyles } from './components/switch/styles';
-export type {
-  SwitchProps,
-  SwitchProviderProps,
-} from './components/switch/types';
+export type { SwitchProps } from './components/switch/types';
 export { Tabs } from './components/tabs';
 export type {
   TabListProps,
@@ -213,16 +211,10 @@ export {
   TextAreaFieldContext,
 } from './components/text-area-field';
 export { TextAreaStyles } from './components/text-area-field/styles';
-export type {
-  TextAreaFieldProps,
-  TextAreaFieldProviderProps,
-} from './components/text-area-field/types';
+export type { TextAreaFieldProps } from './components/text-area-field/types';
 export { TextField, TextFieldContext } from './components/text-field';
 export { TextFieldStyles } from './components/text-field/styles';
-export type {
-  TextFieldProps,
-  TextFieldProviderProps,
-} from './components/text-field/types';
+export type { TextFieldProps } from './components/text-field/types';
 export { Tooltip, TooltipBody, TooltipTrigger } from './components/tooltip';
 export type {
   TooltipBodyProps,
@@ -233,10 +225,12 @@ export { containsExactChildren, expectsIconWrapper } from './lib/react';
 export type {
   ChildrenRenderProps,
   ClassNameRenderProps,
+  ProviderProps,
   RenderProps,
   RenderPropsChildren,
   RenderPropsClassName,
   RenderPropsStyle,
+  SlottedValue,
   StylePropRenderProps,
   StyleRenderProps,
 } from './lib/types';